Move core_grades_get_grades into seperate class
